Change HEAP_FLAG from LSB to MSB
This saves us a from doing a shift when constructing or accessing an inline value. There's an additional rotate when constructing a heap value, but that is an infrequent operation, dominated by the allocation itself. This also has the nice property that all inline values have the same representation as u64.